+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## 2.11.0 - 2024-07-12
 ### Added
 - Support for `dataType=pdf` in `CadenzaClient#downloadData` and `CadenzaClient#fetchData`
 - `CadenzaClient#fetchObjectInfo`